embed-server/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
astroid==3.0.0
async-timeout==4.0.3
attrs==23.1.0
black==23.9.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
dataclasses-json==0.6.1
dill==0.3.7
fastapi==0.103.2
filelock==3.12.4
flake8==6.1.0
flake8-unused-arguments==0.0.13
frozenlist==1.4.0
fsspec==2023.9.2
greenlet==3.0.0
h11==0.14.0
huggingface-hub==0.17.3
idna==3.4
InstructorEmbedding==1.0.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langsmith==0.0.43
MarkupSafe==2.1.3
marshmallow==3.20.1
mccabe==0.7.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numpy==1.26.0
packaging==23.2
pathspec==0.11.2
Pillow==10.0.1
platformdirs==3.11.0
pycodestyle==2.11.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyflakes==3.1.0
pylint==3.0.1
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.14.1
tomlkit==0.12.1
torchvision==0.16.0
tqdm==4.66.1
transformers==4.34.0
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.23.2
yarl==1.9.2
